Final Assignment ST 1
Code all worked fine yesterday but showing some errors today now with the beautiful_soup Parse steps. Not sure why that is. I also tried to save this so that the graph pictures show up in Question 5 and 6 so I hope that works. I have submitted the screenshots showing the images as they ran in the exercise. Thanks!